// typechain-types/RumbleCourt.ts
export declare namespace RumbleCourt {
  export type CaseStruct = {
    caseId: BigNumberish;
    plaintiff: AddressLike;
    caseTitle: string;
    evidenceHash: string;
    filedAt: BigNumberish;
    status: BigNumberish;
  };

  export type CaseStructOutput = [
    caseId: bigint,
    plaintiff: string,
    caseTitle: string,
    evidenceHash: string,
    filedAt: bigint,
    status: bigint
  ] & {
    caseId: bigint;
    plaintiff: string;
    caseTitle: string;
    evidenceHash: string;
    filedAt: bigint;
    status: bigint;
  };

  export type VerdictStruct = {
    caseId: BigNumberish;
    verdictType: BigNumberish;
    reasoning: string;
    timestamp: BigNumberish;
    isFinal: boolean;
  };

  export type VerdictStructOutput = [
    caseId: bigint,
    verdictType: bigint,
    reasoning: string,
    timestamp: bigint,
    isFinal: boolean
  ] & {
    caseId: bigint;
    verdictType: bigint;
    reasoning: string;
    timestamp: bigint;
    isFinal: boolean;
  };
}